fix(`graphviz-dot-indent-width`)!: default from `standard-indent`
What
===

Use `standard-indent` instead of `tab-width` as the default value for `graphviz-dot-indent-width`.

Why
===

The `tab-width` documentation states the following.

> Distance between tab stops (for display of tab characters), in columns.
>
> This controls the width of a TAB character on display.
> The value should be a positive integer.
> Note that this variable doesn't necessarily affect the size of the
> indentation step.  However, if the major mode's indentation facility
> inserts one or more TAB characters, this variable will affect the
> indentation step as well, even if indent-tabs-mode is non-nil.

So when indentation is not done with hard tabs, the `tab-width` should not affect the number of indentations, and therefore I don't think it is correct to treat it as the default value for indentation.

I am not completely confident about what to use then, but I think `standard-indent` is better.

What we are not doing.
===

When I converted texi files to HTML, I did not update the HTML because the tool here is too new or a considerable number of conversions were made.